fix: 修复H2 PostgreSQL模式下DOUBLE类型错误
- DOUBLE改为DOUBLE PRECISION匹配PostgreSQL语法
- 验证通过: 5个测试PASSED, 5个SKIPPED

+++ evcs-order/src/test/resources/schema-h2.sql
@@ -15,7 +15,7 @@ CREATE TABLE charging_order (
     user_id BIGINT,
     start_time TIMESTAMP,
     end_time TIMESTAMP,
-    energy DOUBLE,
+    energy DOUBLE PRECISION,
     duration BIGINT,
     amount DECIMAL(10,2),
     billing_plan_id BIGINT,
